requirements.txt
numpy==1.23.4
torch==1.11.0+cu113
opencv-python==4.6.0.66
scipy==1.9.3
tqdm==4.64.1
torchvision==0.12.0+cu113
fastai==1.0.51
matplotlib==3.6.1
spacy==3.4.2
fastprogress==1.0.3
Pillow==9.1.0
nbconvert==7.2.3
ipython==8.6.0
typing==3.7.4.3
nbformat==5.7.0
traitlets==5.5.0
PyYAML==6.0
pandas==1.5.1
requests==2.28.1
ipywidgets==8.0.2
beautifulsoup4==4.11.1
tensorboardX==1.6
DeOldify==0.0.1
future==0.18.2
customtkinter==4.6.3